+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 0.11.2 (Unreleased)
 -------------------
 - Enh: Automated code refactoring for HumHub 1.18 using Rector
+- Enh #232: Add custom notification endpoint
 
 0.11.1 (November 19, 2025)
 --------------------------